refactor(front): del comments, new file for script
